+++ src/CategorySource.php
@@ -17,15 +17,18 @@ final class CategorySource
      * @param string $name Category name.
      * @param MessageReaderInterface $reader Message reader to get messages from for this category.
      * @param MessageFormatterInterface|null $formatter Message formatter to format messages with for this category.
+     * @param MessageWriterInterface|null $writer Message writer to write messages for this category.
      */
     public function __construct(
         string $name,
         private MessageReaderInterface $reader,
-        private ?MessageFormatterInterface $formatter = null
+        private ?MessageFormatterInterface $formatter = null,
+        private ?MessageWriterInterface $writer = null,
     ) {
         if (!preg_match('/^[a-z0-9_-]+$/i', $name)) {
             throw new RuntimeException('Category name is invalid. Only letters and numbers are allowed.');
         }
+
         $this->name = $name;
     }
 
@@ -78,6 +81,38 @@ final class CategorySource
     public function getMessages(string $locale): array
     {
         return $this->reader->getMessages($this->name, $locale);
+    }
+
+    /**
+     * Writes a set of messages for a specified category and locale.
+     *
+     * @psalm-param array<string, array<string, string>> $messages
+     *
+     * @param string $locale Locale to write messages for.
+     * @param array $messages A set of messages to write. The format is the following:
+     * ```php
+     * [
+     *   'key1' => [
+     *     'message' => 'translation1',
+     *     // Extra metadata that writer may use:
+     *     'comment' => 'Translate carefully!',
+     *   ],
+     *   'key2' => [
+     *     'message' => 'translation2',
+     *     // Extra metadata that writer may use:
+     *     'comment' => 'Translate carefully!',
+     *   ],
+     * ]
+     * ```
+     *
+     * @throws UnwritableCategorySourceException When $write is not configured, or it's impossible to write the messages into the source.
+     */
+    public function write(string $locale, array $messages): void
+    {
+        if ($this->writer === null) {
+            throw new UnwritableCategorySourceException($this->name);
+        }
+        $this->writer->write($this->name, $locale, $messages);
     }
 
     /**